refactor: Sequence.lua now allows for plugins. First one being ButtonClick.
